Refactor order mapping logic into dedicated classes for improved maintainability
